+++ src/main.py
@@ -1,8 +1,0 @@
-# __main__.py
-from .main import main
-
-if __name__ == "__main__":
-    main()
-
-# now you can run:
-# $ python -m nespressobot